ispravka sale,profila ak i lekara

+++ src/main/java/rest/controller/SalaController.java
@@ -16,9 +16,14 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import rest.domain.Klinika;
+import rest.domain.Pregled;
 import rest.domain.Sala;
 import rest.domain.User;
 import rest.dto.SalaDTO;
+import rest.pk.SalaPK;
+import rest.service.KlinikaService;
+import rest.service.PregledService;
 import rest.service.SalaService;
 
 
@@ -27,26 +32,30 @@ import rest.service.SalaService;
 @RequestMapping("rest/sala")
 public class SalaController {
 	@Autowired
-	
 	private SalaService salaService;
-
+	@Autowired
+	private KlinikaService klinikaService;
+	@Autowired
+	private PregledService pregledService;
 	@GetMapping
 	public ResponseEntity<List<SalaDTO>> getSala() {
-		
+		System.out.println("prvi maj");
 		List<Sala> sale = salaService.findAll();
-
+		System.out.println("1.maj: "+sale.isEmpty());
 		List<SalaDTO> saleDTO = new ArrayList<>();
 		for (Sala s : sale) {
-			saleDTO.add(new SalaDTO(s));
+			saleDTO.add(new SalaDTO(s.getId(),s.getKlinika().getId(),s.getNaziv()));
 		}
-
+		
+		System.out.println("top");
 		return new ResponseEntity<>(saleDTO, HttpStatus.OK);
 	}
 	@PutMapping(value="/izmeni",consumes = "application/json")
 	public ResponseEntity<SalaDTO> updateCourse(@RequestBody SalaDTO salaDTO) {
 
-		// a course must exist
-		Sala sala = salaService.findOne(salaDTO.getId());
+		// a course must exis
+		SalaPK pk=new SalaPK(salaDTO.getBrojSale(),salaDTO.getKlinika());
+		Sala sala = salaService.findOne(pk);
 
 		if (sala == null) {
 			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
@@ -55,23 +64,37 @@ public class SalaController {
 		sala.setNaziv(salaDTO.getNaziv());
 
 		sala = salaService.save(sala);
-		return new ResponseEntity<>(new SalaDTO(sala), HttpStatus.OK);
+		return new ResponseEntity<>(new SalaDTO(sala.getId(),sala.getKlinika().getId(),sala.getNaziv()), HttpStatus.OK);
 	}
-	@DeleteMapping(value = "/{id}")
-	public ResponseEntity<Void> deleteCourse(@PathVariable Integer id) {
-
+	@PostMapping(value = "/obrisi")
+	public ResponseEntity<Void> deleteCourse(@RequestBody SalaPK id) {
 		Sala sala = salaService.findOne(id);
 		System.out.println("brisanje");
-		if (sala != null) {
-			salaService.remove(id);
-			return new ResponseEntity<>(HttpStatus.OK);
-		} else {
+		List<Pregled> ztermini = pregledService.findZauzete(sala);
+		System.out.println(ztermini.isEmpty());
+		if (ztermini.isEmpty()){
+			System.out.println("brisanje");
+			if (sala != null) {
+				salaService.remove(id);
+				return new ResponseEntity<>(HttpStatus.OK);
+			} else {
+				return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+			}
+		}else {
 			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 		}
 	}
 	@PostMapping(value="/dodaj",consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<User> login(@RequestBody SalaDTO salaDTO){
-		Sala sala= new Sala(salaDTO.getId(),salaDTO.getNaziv());
+	public ResponseEntity<User> dodajSalu(@RequestBody SalaDTO salaDTO){
+		//SalaPK pk=new SalaPK(salaDTO.getBrojSale(),salaDTO.getKlinika());
+		Klinika k= klinikaService.findOne(salaDTO.getKlinika());
+		/*SalaPK pk=new SalaPK(salaDTO.getBrojSale(),salaDTO.getKlinika());
+		Sala salab = salaService.findOne(pk);
+		if (salab != null) {
+			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+		}*/
+		
+		Sala sala= new Sala(k,salaDTO.getBrojSale(),salaDTO.getNaziv());
 		salaService.save(sala);
 		return new ResponseEntity<>(HttpStatus.OK);
 	}
